Refactor release meta-data into a Release struct
Now the running and name variables are part of a Release struct which
also includes a timestamp for the time when the release started.  This
makes handling the meta-data a bit simpler because it's all in one
place.  During this refactor I also rewrote the /start handler's log
strings to include time and author when available.