Dockerise the app
Add a Dockerfile to build a releasestatus container from scratch.
It adds the Linux-built releasestatus binary as the Docker command, as
well as a zoneinfo archive for setting the timezone inside the
container.

It has a Makefile to reduce the amount of typing for the Docker build
and a .gitignore file to ignore the binaries and archive used during the
process.

The Makefile's default target will build and run the Docker image.  It
has some default variables which can be overridden from the environment:

  - TZ:   the timezone the container will run in (default: CET)
  - PORT: the port the service will run on (default: 8080)
  - TAG:  the image tag in Docker (default: releasestatus)

It also has separate make targets:

  - container: just build the image
  - run: run the image (assumes it is built)
  - clean: removes the binaries and archive used during the process

The container from scratch is very small, it contains only the go binary
and zone info.  For this to work correctly the binary is built with
netgo included (which has a native implementation of host lookup etc),
so that it can handle network requests.  The go build tool uses special
environment variables to build for Linux amd64, suitable for running in
Docker.  Because I also want to be able to set the time zone for the
container (I'd like the logging timestamps to be in Central European
Time) it needs a zoneinfo archive, which is borrowed from the system's
/usr/share/zoneinfo directory at build time.  These are the only 2 files
the container currently needs to run, and in fact, if you're fine with
running it in UCT, you don't even need the zoneinfo file.

The run target will use your environment variables (if provided) to
override my default settings, and the container target can use the TAG
variable if you'd like to give the image a special tag for deployment.
The container will be removed when it stops running, to avoid clutter
among your containers.

This all-together makes building, running and deploying the dockerised
version of this service really convenient and easy.