Add a block endpoint which stops new releases from being started
When we need to keep the test site stable for a while it's useful to be
able to block new releases even when no release is running.  To support
this, the new block endpoint will toggle a new Release field, Blocked.

A release which is blocked does not allow a new release to start and
will persist the blocked state when a running release is stopped.